+++ src/packages/components/src/components/editors/SelectBoxComponent.tsx
@@ -1,4 +1,4 @@
-import { debounce } from 'lodash'
+import debounce from 'lodash/debounce'
 import { CircleX } from 'lucide-react'
 import { ChangeEvent, FC, useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { FieldContainer } from './FieldContainer'